Add Equipment Add/Edit/Remove; Add Building Add/Edit/Remove

+++ RoomReservation.Implementation/Repositories/RepositoryGenericBase.cs
@@ -17,7 +17,7 @@ namespace RoomReservation.Implementation.Repositories {
             Logger = logger;
         }
 
-        public async Task<T?> GetOneAsync(long key, params Expression<Func<T, object>>[] includes)
+        public async Task<T?> GetOneAsync(int key, params Expression<Func<T, object>>[] includes)
         {
             try
             {
